# openapi.yaml
openapi: 3.0.3
info:
  title: Capitok Gateway API
  version: 0.1.0
  description: API for archiving raw agent conversations and querying derived recall records.
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service status
  /v1/ingest:
    post:
      summary: Archive one interaction and enqueue derived recall indexing
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Raw interaction archived and derived indexing accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '401':
          description: Missing or invalid API key
        '403':
          description: Missing scope
        '422':
          description: Validation error
  /v1/search:
    get:
      summary: Search derived recall records
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: top_k
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Derived recall search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '401':
          description: Missing or invalid API key
        '403':
          description: Missing scope
        '422':
          description: Validation error
  /v1/sessions:
    get:
      summary: List archived sessions or raw records
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: view
          required: false
          schema:
            type: string
            enum: [sessions, records]
            default: sessions
        - in: query
          name: source
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Session list or raw record view
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401':
          description: Missing or invalid API key
        '403':
          description: Missing scope
        '422':
          description: Validation error
  /v1/sessions/{session_id}:
    get:
      summary: Get one archived session timeline
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
        - in: query
          name: source
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: Session timeline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
        '401':
          description: Missing or invalid API key
        '403':
          description: Missing scope
        '409':
          description: Multiple sources share the same session_id; pass source
        '404':
          description: Session not found
        '422':
          description: Validation error

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    IngestRequest:
      type: object
      required:
        - session_id
        - user_id
      properties:
        session_id:
          type: string
        user_id:
          type: string
        source:
          type: string
          default: agent
        input:
          type: string
        output:
          type: string
        metadata:
          type: object
          additionalProperties: true
    IngestResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string
    SearchResult:
      type: object
      required:
        - id
        - session_id
        - user_id
        - text
        - score
        - created_at
      properties:
        id:
          type: string
        session_id:
          type: string
        user_id:
          type: string
        text:
          type: string
        score:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
    SearchResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
    SessionListItem:
      type: object
      required:
        - session_id
        - source
        - preview
        - started_at
        - updated_at
        - record_count
      properties:
        session_id:
          type: string
        source:
          type: string
        preview:
          type: string
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        record_count:
          type: integer
    SessionRecordListItem:
      type: object
      required:
        - id
        - session_id
        - source
        - created_at
        - input
        - output
      properties:
        id:
          type: string
        session_id:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
        input:
          type: string
        output:
          type: string
    SessionTimelineItem:
      type: object
      required:
        - id
        - created_at
        - input
        - output
        - metadata
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        input:
          type: string
        output:
          type: string
        metadata:
          type: object
          additionalProperties: true
    SessionListResponse:
      type: object
      required:
        - view
        - items
      properties:
        view:
          type: string
          enum: [sessions, records]
        items:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/SessionListItem'
              - $ref: '#/components/schemas/SessionRecordListItem'
    SessionDetailResponse:
      type: object
      required:
        - session_id
        - source
        - started_at
        - updated_at
        - record_count
        - items
      properties:
        session_id:
          type: string
        source:
          type: string
        started_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        record_count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionTimelineItem'